Split go tests
